create Uploader class and moved to that of Photo class
