+++ tests/test_photo.py
@@ -1,12 +1,9 @@
 from datetime import datetime
 from pathlib import Path
 
-import boto3
 import pytest
-from PIL import Image
-
-import photo_backup.consts as consts
 from photo_backup.photo import Photo
+from PIL import Image
 
 
 class TestPhotoInit:
@@ -45,19 +42,3 @@ class TestPhotoCreateDstPath:
         actual = Photo("data/images/IMG_9235.jpeg").create_dst_path()
         expected = "2020/08/20200813140708.jpeg"
         assert actual == expected
-
-
-class TestPhotoUploadToS3:
-    def test_upload(self):
-        Photo("data/images/IMG_9235.jpeg").upload()
-
-        s3 = boto3.client("s3")
-        response = s3.list_objects_v2(
-            Bucket=consts.S3_BUCKET, Prefix="2020/08"
-        )
-        uploaded_filename = response["Contents"][0]["Key"]
-        expected = "2020/08/20200813140708.jpeg"
-
-        assert uploaded_filename == expected
-
-        s3.delete_object(Bucket=consts.S3_BUCKET, Key=expected)
